Improve Contacts

+++ src/mySimpleMessageService.Application/Contacts/Handlers/DeleteContactCommandHandler.cs
@@ -2,6 +2,7 @@
 using System.Threading.Tasks;
 using MediatR;
 using mySimpleMessageService.Application.Contacts.Commands;
+using mySimpleMessageService.Application.Exceptions;
 using mySimpleMessageService.Application.Interfaces;
 
 namespace mySimpleMessageService.Application.Contacts.Handlers
@@ -18,6 +19,9 @@ namespace mySimpleMessageService.Application.Contacts.Handlers
 
         public async Task<Unit> Handle(DeleteContactCommand request, CancellationToken cancellationToken)
         {
+            var contact = _contactRepository.GetAsync(request.Id);
+            if (contact is null)
+                throw new ContactNotFoundException(request.Id);
             await _contactRepository.DeleteAsync(request.Id);
             return Unit.Value;
         }
